modificaciones del profe

+++ backend/components/users/user_controller.js
@@ -7,23 +7,31 @@ export class UserController {
   }
 
   async get(req, res) {
-    const userList = await this.userService.getList();
+    checkPermission(req, ['admin']);
+    const { uuid } = req.params;
+    const filters = {};
+    if (uuid) {
+      filters.uuid = uuid;
+    } 
+    const userList = await this.userService.getList(filters);
     res.send(userList);
   }
 
   async post(req, res) {
+    checkPermission(req, ['admin']);
     await this.userService.create(req.body);
     res.status(204).end();
   }
 
   async patch(req, res) {
+    checkPermission(req, ['admin']);
     const { uuid } = req.params;
     await this.userService.update(uuid, req.body);
     res.status(200).end();
   }
 
   async delete(req,res){
-    //checkPermission(req, ['admin']);
+    checkPermission(req, ['admin']);
     await this.userService.deleteForUuid(req.params.uuid);
     res.status(204).end();
   }
